# utils/swagger-new.yaml
openapi: 3.0.0
info:
  title: Smart Contract API - Online Voting System
  description: API for interacting with an election smart contract on the blockchain
  version: 1.0.0
servers:
  - url: http://localhost:3001
    description: Local development server
  - url: https://evoting-p0zm.onrender.com
    description: Cloud server old
  - url: https://evoting-backend-0v72.onrender.com
    description: Cloud server New
paths:
...
  /api/register:
    post:
      tags:
        - Admin
      summary: Register a candidate (Multi-Election Type Support)
      description: Register a new candidate for an election with support for National Assembly, Gup, Mangmi, and Tshogpa elections (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                electionId:
                  type: string
                  description: ID of the election
                  example: "NA_2025"
                candidate:
                  type: string
                  description: Name of the candidate
                  example: "Tshering Dorji"
                location:
                  type: object
                  description: Geographical location (NEW FORMAT - supports all election types)
                  properties:
                    dzongkhag:
                      type: string
                      description: Dzongkhag name
                      example: "Thimphu"
                    demkhong:
                      type: string
                      description: Demkhong name (for National Assembly)
                      example: "Thimphu"
                    gewog:
                      type: string
                      description: Gewog name (for Gup/Mangmi/C_Tshogpa)
                      example: "Chang"
                    chiwog:
                      type: string
                      description: Chiwog name (for C_Tshogpa only)
                      example: "Babesa"
                    thromde:
                      type: string
                      description: Thromde/Municipality name (for THROMPOEN and T_TSHOGPA)
                      example: "Thimphu Thromde"
                electionType:
                  type: string
                  description: Type of election (optional - can be auto-detected, case-insensitive)
                  enum:
                    [
                      NA,
                      NC,
                      GUP,
                      MANGMI,
                      C_TSHOGPA,
                      THROMPOEN,
                      T_TSHOGPA,
                    ]
                  example: "NA"
                demkhong:
                  type: string
                  description: Legacy format - demkhong only (still supported for backward compatibility)
                  example: "Thimphu"
              required:
                - electionId
                - candidate
            examples:
              nationalAssembly:
                summary: National Assembly Election (NA)
                value:
                  electionId: "NA_2025"
                  candidate: "Tshering Dorji"
                  location:
                    dzongkhag: "Thimphu"
                    demkhong: "Thimphu"
                  electionType: "NA"
              nationalCouncil:
                summary: National Council Election (NC)
                value:
                  electionId: "NC_2025"
                  candidate: "Sonam Phuntsho"
                  location:
                    dzongkhag: "Paro"
                  electionType: "NC"
              gupElection:
                summary: Gup Election (GUP)
                value:
                  electionId: "GUP_2025_THIMPHU"
                  candidate: "Karma Wangchuk"
                  location:
                    dzongkhag: "Paro"
                    gewog: "Lungnyi"
                  electionType: "GUP"
              mangmiElection:
                summary: Mangmi Election (MANGMI)
                value:
                  electionId: "MANGMI_2025_PARO"
                  candidate: "Dorji Wangmo"
                  location:
                    dzongkhag: "Paro"
                    gewog: "Lungnyi"
                  electionType: "MANGMI"
              chiwogTshogpaElection:
                summary: Chiwog Tshogpa Election (C_TSHOGPA)
                value:
                  electionId: "C_TSHOGPA_2025"
                  candidate: "Pema Thinley"
                  location:
                    dzongkhag: "Paro"
                    gewog: "Lungnyi"
                    chiwog: "Tsento"
                  electionType: "C_TSHOGPA"
              thromponElection:
                summary: Thrompon Election (THROMPOEN)
                value:
                  electionId: "THROM_2025"
                  candidate: "Ugyen Dorji"
                  location:
                    dzongkhag: "Thimphu"
                    thromde: "Thimphu Thromde"
                  electionType: "THROMPOEN"
              thromponTshogpaElection:
                summary: Thrompon Tshogpa Election (T_TSHOGPA)
                value:
                  electionId: "T_TSHOGPA_2025"
                  candidate: "Dawa Tshering"
                  location:
                    dzongkhag: "Thimphu"
                    thromde: "Phuentsholing Thromde"
                  electionType: "T_TSHOGPA"
              legacyFormat:
                summary: Legacy Format (Backward Compatible)
                value:
                  electionId: "NA_2025"
                  candidate: "Tshering Dorji"
                  demkhong: "Thimphu"
      responses:
        "200":
          description: Candidate registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Candidate registered successfully"
                  candidate:
                    type: string
                    example: "Tshering Dorji"
                  electionType:
                    type: string
                    example: "NA"
                  location:
                    type: string
                    example: "Thimphu/Thimphu"
                  txHash:
                    type: string
                    example: "0xabc123..."
        "400":
          description: Bad request - missing or invalid fields
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  missingFields:
                    type: array
                    items:
                      type: string
                  requiredFields:
                    type: array
                    items:
                      type: string
        "409":
          description: Candidate already registered
        "403":
          description: Unauthorized - only contract owner can register
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /api/remove:
    delete:
      tags:
        - Admin
      summary: Remove a candidate
      description: Remove a candidate from an election (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                electionId:
                  type: string
                  description: ID of the election
                candidate:
                  type: string
                  description: Address or identifier of the candidate
              required:
                - electionId
                - candidate
      responses:
        "200":
          description: Candidate removed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  txHash:
                    type: string
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /api/vote:
    post:
      tags:
        - Voting
      summary: Cast a vote
      description: Cast a vote in an election with polling station tracking (works for all election types)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                electionId:
                  type: string
                  description: ID of the election
                  example: "GUP_2025_THIMPHU"
                uid:
                  type: string
                  description: Unique identifier of the voter (CID)
                  example: "11234567890123"
                candidate:
                  type: string
                  description: Name of the candidate being voted for
                  example: "Karma Wangchuk"
                gender:
                  type: string
                  description: Gender of the voter
                  enum: [Male, Female]
                  example: "Male"
                pollingStation:
                  type: string
                  description: Polling station identifier where the vote was cast
                  example: "PS_Thimphu_01"
              required:
                - electionId
                - uid
                - candidate
                - gender
                - pollingStation
      responses:
        "200":
          description: Vote cast successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Vote cast successfully"
                  txHash:
                    type: string
                    example: "0xdef456..."
                    txStatus:
                    type: string
                    example: "success"
                  pollingStation:
                    type: string
                    example: "PS_Thimphu_01"
                  explorerLink:
                    type: string
                    example: "https://amoy.polygonscan.com/tx/0xdef456..."
        "400":
          description: Bad request - missing required fields
        "404":
          description: Election or candidate not found
        "409":
          description: Already voted in this election
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /api/end:
    post:
      tags:
        - Admin
      summary: End an election
      description: Mark an election as ended (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                electionId:
                  type: string
                  description: ID of the election to end
              required:
                - electionId
      responses:
        "200":
          description: Election ended successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  txHash:
                    type: string
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT